+++ blocks/table/src/components/ButtonField/index.tsx
@@ -2,7 +2,7 @@ import { useBlock } from '@appsemble/preact';
 import { Icon } from '@appsemble/preact-components';
 import classNames from 'classnames';
 import { Fragment, VNode } from 'preact';
-import { useCallback } from 'preact/hooks';
+import { useCallback, useMemo } from 'preact/hooks';
 
 import { Button as ButtonFieldType } from '../../../block';
 
@@ -37,6 +37,7 @@ export function ButtonField({
       rounded,
       size = 'normal',
       title,
+      ...button
     },
     ...field
   },
@@ -49,6 +50,10 @@ export function ButtonField({
     utils: { remap },
   } = useBlock();
 
+  const disabled = useMemo(
+    () => Boolean(remap(button.disabled, item, { index, repeatedIndex })),
+    [button.disabled, index, item, remap, repeatedIndex],
+  );
   const action = actions[field.onClick];
   const className = classNames('button', `is-${size}`, {
     'is-rounded': rounded,
@@ -68,15 +73,31 @@ export function ButtonField({
   );
 
   const onClick = useCallback(() => {
+    if (disabled) {
+      return;
+    }
+
     action(item);
-  }, [action, item]);
+  }, [action, disabled, item]);
 
   return action?.type === 'link' ? (
-    <a className={className} href={action.href()} onClick={onClick} title={remappedTitle}>
+    <a
+      className={className}
+      disabled={disabled}
+      href={action.href()}
+      onClick={onClick}
+      title={remappedTitle}
+    >
       {content}
     </a>
   ) : (
-    <button className={className} onClick={onClick} title={remappedTitle} type="button">
+    <button
+      className={className}
+      disabled={disabled}
+      onClick={onClick}
+      title={remappedTitle}
+      type="button"
+    >
       {content}
     </button>
   );