+++ packages/server/src/utils/resource.ts
@@ -34,9 +34,14 @@ import { sendNotification, SendNotificationOptions } from './sendNotification';
  *
  * @param app - The app to get the resource definition of
  * @param resourceType - The name of the resource definition to get.
+ * @param view - The view that’s being used.
  * @returns The matching resource definition.
  */
-export function getResourceDefinition(app: App, resourceType: string): ResourceDefinition {
+export function getResourceDefinition(
+  app: App,
+  resourceType: string,
+  view?: string,
+): ResourceDefinition {
   if (!app) {
     throw notFound('App not found');
   }
@@ -45,6 +50,10 @@ export function getResourceDefinition(app: App, resourceType: string): ResourceD
 
   if (!definition) {
     throw notFound(`App does not have resources called ${resourceType}`);
+  }
+
+  if (view && !definition.views[view]) {
+    throw notFound(`View ${view} does not exist for resource type ${resourceType}`);
   }
 
   return definition;
@@ -356,6 +365,7 @@ export function processResourceBody(
     {
       base: '#',
       preValidateProperty,
+      nestedErrors: true,
       rewrite(value, { format }, options, { path }) {
         if (
           Array.isArray(resource)